added spark max motor type and started adding enums.
