+++ codeGenerator/src/robotConfiguration/ApplicationData.csproj
@@ -42,6 +42,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="applicationData_MotorControllers.cs" />
     <Compile Include="testApplicationData.cs" />
     <Compile Include="applicationDataConfig.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
